renamed a couple files, fixed pep8 conformance to pass tests. Fixed broken generation of pep257 files in fileTests.py that was making it impossible to add directories to _pep8Files
