+++ DRP/tests/ModelBuildingTests/__init__.py
@@ -1,15 +1,15 @@
 '''This package provides tests model building in the DRP project'''
 import unittest
-import ModelContainer
-import BCR
 import J48
 import KNN
 import NaiveBayes
+import SVM_basic
+import SVM_BCR
 
 suite = unittest.TestSuite([
-    ModelContainer.suite,
-    BCR.suite,
-    KNN.suite,
     J48.suite,
+    KNN.suite,
     NaiveBayes.suite,
+    SVM_basic.suite,
+    SVM_BCR.suite,
 ])
